Compiler compatibility, other fixes (#1188)
* add missing data

* fix return value

* d_camera fix

* missing GXEnd

* struct fixes

* fix include

+++ src/d/d_map.cpp
@@ -345,9 +345,9 @@ dStage_FloorInfo_dt_c* dMap_GetFloorInfoDtPFromFloorNo(dStage_FloorInfo_c* floor
 }
 
 /* 80045AEC-80045C24       .text dMap_GetFloorNoForDmap__FP11dStage_dt_cif */
-int dMap_GetFloorNoForDmap(dStage_dt_c* stag, int param_2, f32 param_3) {
+u8 dMap_GetFloorNoForDmap(dStage_dt_c* stag, int param_2, f32 param_3) {
     dStage_FloorInfo_c* floor = stag->getFloorInfo();
-    int floorNo = Floor_1F;
+    u8 floorNo = Floor_1F;
     dStage_FloorInfo_dt_c* floorData = dMap_GetFloorInfoDtP(floor, param_3);
 
     if (floorData) {
